requirements.txt
uvicorn==0.20.0
fastapi==0.92.0
keras
requests
tensorflow
Jinja2
python-multipart
Pillow